Update bug_report.md
I have an issue in SL where when I am passively bombed by my enemies my city regens, but i need redstone in my trading ships so that the pistons can't extend and drop the crates, the issue is that redstone does not regen, instead it just drops as if it were blew up by TNT in vanilla, I cannot trade as Orlandia is bombed a lot. Whenever I am bombed I got to decomp my ships, and then rebuild them. To reproduce: 1. Fly to a claimed area. 2. Build a structure with redstone. 3. Shoot at it, you will see that whatever the structure is made of will regen, but most of the redstone will drop as an item and not regen, as the block under it blew up updating the floating redstone making it break. Whatever the structure is made of will regen. I do not have any screenshots or videos, but I explained it as thoroughly as I possibly can. This is an important error as it is likely what is causing the lack of trading in the game.